Refactored non-browser tests to use CubanoDemoFixture
