+++ src/test/java/example/CubanoDemoBrowserFixture.java
@@ -1,25 +1,24 @@
 package example;
 
+import example.cubanohttpeasy.HttpEasyConfigurator;
 import org.concordion.api.*;
 import org.concordion.api.extension.Extension;
-import org.concordion.cubano.config.Config;
-import org.concordion.cubano.config.ProxyConfig;
 import org.concordion.cubano.driver.concordion.ExceptionHtmlCaptureExtension;
-import org.concordion.cubano.driver.http.HttpEasy;
 import org.concordion.cubano.framework.ConcordionBrowserFixture;
 import org.concordion.cubano.framework.resource.CloseListener;
 import org.concordion.cubano.framework.resource.ResourceScope;
-import org.concordion.cubano.template.AppConfig;
 import org.concordion.slf4j.ext.ReportLogger;
 import org.concordion.slf4j.ext.ReportLoggerFactory;
 
 import java.io.Closeable;
-import java.net.InetSocketAddress;
-import java.net.Proxy;
 
 /**
- * Customises the test specification and provides some helper methods
- * so the tests can access the storyboard, browser, etc.
+ * A base class for extension by fixtures that invoke a browser, and may also use HttpEasy.
+ *
+ * Customises the test specification and provides some helper methods so the tests can access the storyboard, browser, etc.
+ *
+ * @see CubanoDemoIndex for fixtures that don't contain assertions
+ * @see CubanoDemoFixture for fixtures that don't invoke a browser
  */
 @ConcordionResources("/customConcordion.css")
 @FailFast
@@ -30,26 +29,10 @@ public abstract class CubanoDemoBrowserFixture extends ConcordionBrowserFixture 
     private final ExceptionHtmlCaptureExtension htmlCapture = new ExceptionHtmlCaptureExtension(getStoryboard(), getBrowser());
 
     static {
-        ProxyConfig proxyConfig = Config.getInstance().getProxyConfig();
-        AppConfig config = AppConfig.getInstance();
-        config.logSettings();
-
-        // Set the proxy rules for all rest requests made during the test run
-        HttpEasy.withDefaults().allowAllHosts().trustAllCertificates().baseUrl(config.getBaseUrl());
-
-        if (proxyConfig.isProxyRequired()) {
-            HttpEasy.withDefaults()
-                    .proxy(new Proxy(Proxy.Type.HTTP,
-                            new InetSocketAddress(proxyConfig.getProxyHost(), proxyConfig.getProxyPort())))
-                    .bypassProxyForLocalAddresses(true);
-
-            if (!proxyConfig.getProxyUsername().isEmpty() && !proxyConfig.getProxyPassword().isEmpty()) {
-                HttpEasy.withDefaults().proxyAuth(proxyConfig.getProxyUsername(), proxyConfig.getProxyPassword());
-            }
-        }
+        HttpEasyConfigurator.applyTrustingConfig();
     }
 
-    /** Override the default reportLogger. **/
+    /** Override the default fixture logger. **/
     public CubanoDemoBrowserFixture() {
         super.withFixtureListener(new CubanoDemoFixtureLogger());
     }
@@ -67,7 +50,6 @@ public abstract class CubanoDemoBrowserFixture extends ConcordionBrowserFixture 
 
             @Override
             public void afterClosing(Closeable resource) {
-                // Shouldn't need to do this, but this event getting triggered AFTER the Storyboard's afterExample event listener.
                 getStoryboard().setAcceptCards(true);
             }
         };
