+++ src/test/java/example/CubanoDemoFixture.java
@@ -1,29 +1,25 @@
 package example;
 
-import ch.qos.logback.classic.Level;
-import org.concordion.api.extension.Extension;
+import example.cubanohttpeasy.HttpEasyConfigurator;
+import org.concordion.api.ConcordionResources;
+import org.concordion.api.FailFast;
 import org.concordion.cubano.framework.ConcordionFixture;
-import org.concordion.ext.LogbackLogMessenger;
-import org.concordion.ext.LoggingTooltipExtension;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
+ * A base class for extension by fixtures that contain assertions.
+ *
+ * @see CubanoDemoIndex for fixtures that don't contain assertions
+ * @see CubanoDemoBrowserFixture for fixtures that invoke a browser
  */
+@ConcordionResources("/customConcordion.css")
+@FailFast
 public abstract class CubanoDemoFixture extends ConcordionFixture {
-    private final Logger tooltipLogger = LoggerFactory.getLogger("TOOLTIP_" + this.getClass().getName());
-
-    @Extension
-    private final LoggingTooltipExtension tooltipExtension = new LoggingTooltipExtension(new LogbackLogMessenger(tooltipLogger.getName(), Level.ALL, true, "%msg%n"));
+    static {
+        HttpEasyConfigurator.applyTrustingConfig();
+    }
 
     /** Override the default logger. **/
     public CubanoDemoFixture() {
         super.withFixtureListener(new CubanoDemoFixtureLogger());
-    }
-
-    public void addConcordionTooltip(final String message) {
-        // Logging at debug level means the message won't make it to the console, but will make
-        // it to the logs (based on included logback configuration files)
-        tooltipLogger.debug(message);
     }
 }
