+++ src/test/java/example/service/RestRequestFixture.java
@@ -1,11 +1,11 @@
 package example.service;
 
-import example.CubanoDemoBrowserFixture;
+import example.CubanoDemoFixture;
 import org.concordion.cubano.template.driver.services.ExampleRestApi;
 
 import java.io.IOException;
 
-public class RestRequestFixture extends CubanoDemoBrowserFixture {
+public class RestRequestFixture extends CubanoDemoFixture {
 
     public boolean callService() throws IOException {
         return new ExampleRestApi()
